requirements.txt
# Core Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0

# Experiment Tracking
tensorboard>=2.7.0

